Move assigned attribute dataloaders to separate file
